Add magic description (#327)

+++ src/arena/Constuructors/HealMagicConstructor.ts
@@ -73,7 +73,7 @@ export abstract class Heal extends AffectableAction {
     const maxHp = target.stats.val('maxHp');
     const curHp = target.stats.val('hp');
 
-    const allHeal = MiscService.randInt(hl.min, hl.max) * proc;
+    const allHeal = MiscService.randFloat(hl.min, hl.max) * proc;
     const maxHeal = maxHp - curHp;
     const healEffect = Math.min(maxHeal, allHeal);
